Move internal/testtools to testtools
