+++ cmd/gazelle/integration_test.go
@@ -30,8 +30,8 @@ import (
 	"testing"
 
 	"github.com/bazelbuild/bazel-gazelle/config"
-	"github.com/bazelbuild/bazel-gazelle/internal/testtools"
 	"github.com/bazelbuild/bazel-gazelle/internal/wspace"
+	"github.com/bazelbuild/bazel-gazelle/testtools"
 )
 
 // skipIfWorkspaceVisible skips the test if the WORKSPACE file for the
