+++ cmd/move_labels/move_labels_test.go
@@ -19,7 +19,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/bazelbuild/bazel-gazelle/internal/testtools"
+	"github.com/bazelbuild/bazel-gazelle/testtools"
 )
 
 func TestMoveLabels(t *testing.T) {
