+++ language/proto/BUILD.bazel
@@ -48,12 +48,12 @@ go_test(
     deps = [
         "//config:go_default_library",
         "//internal/merger:go_default_library",
-        "//internal/testtools:go_default_library",
         "//label:go_default_library",
         "//language:go_default_library",
         "//repos:go_default_library",
         "//resolve:go_default_library",
         "//rule:go_default_library",
+        "//testtools:go_default_library",
         "//vendor/github.com/bazelbuild/buildtools/build:go_default_library",
         "//walk:go_default_library",
     ],
